[CALCITE-457]: Non-ansi join should push join condition into join operator.
Revise code based on review comments. Add unit test case.
