+++ core/src/test/resources/sql/misc.oq
@@ -303,17 +303,19 @@ EnumerableCalc(expr#0..7=[{inputs}], expr#8=[IS NOT NULL($t5)], expr#9=[NOT($t8)
                 EnumerableTableScan(table=[[hr, emps]])
               EnumerableTableScan(table=[[hr, depts]])
     EnumerableAggregate(group=[{1}], agg#0=[MIN($0)])
-      EnumerableCalc(expr#0..4=[{inputs}], expr#5=[true], expr#6=[90], expr#7=[+($t1, $t6)], expr#8=[CAST($t0):INTEGER NOT NULL], expr#9=[=($t7, $t8)], $f0=[$t5], empid=[$t0], $condition=[$t9])
-        EnumerableJoin(condition=[true], joinType=[inner])
-          EnumerableAggregate(group=[{0}])
-            EnumerableSemiJoin(condition=[=($1, $6)], joinType=[inner])
-              EnumerableTableScan(table=[[hr, emps]])
-              EnumerableCalc(expr#0..4=[{inputs}], expr#5=[true], $f0=[$t5], deptno0=[$t0])
-                EnumerableJoin(condition=[=($0, $1)], joinType=[inner])
-                  EnumerableAggregate(group=[{1}])
-                    EnumerableTableScan(table=[[hr, emps]])
-                  EnumerableTableScan(table=[[hr, depts]])
-          EnumerableTableScan(table=[[hr, depts]])
+      EnumerableCalc(expr#0..6=[{inputs}], expr#7=[true], $f0=[$t7], empid=[$t0])
+        EnumerableJoin(condition=[=($1, $6)], joinType=[inner])
+          EnumerableCalc(expr#0=[{inputs}], expr#1=[CAST($t0):INTEGER NOT NULL], proj#0..1=[{exprs}])
+            EnumerableAggregate(group=[{0}])
+              EnumerableSemiJoin(condition=[=($1, $6)], joinType=[inner])
+                EnumerableTableScan(table=[[hr, emps]])
+                EnumerableCalc(expr#0..4=[{inputs}], expr#5=[true], $f0=[$t5], deptno0=[$t0])
+                  EnumerableJoin(condition=[=($0, $1)], joinType=[inner])
+                    EnumerableAggregate(group=[{1}])
+                      EnumerableTableScan(table=[[hr, emps]])
+                    EnumerableTableScan(table=[[hr, depts]])
+          EnumerableCalc(expr#0..3=[{inputs}], expr#4=[90], expr#5=[+($t0, $t4)], proj#0..3=[{exprs}], $f4=[$t5])
+            EnumerableTableScan(table=[[hr, depts]])
 !plan
 
 # Filter combined with an OR filter.
